requirements.txt
streamlit==1.27.0
pymongo==4.5.0
requests==2.31.0
pandas==2.1.0
plotly==5.17.0
python-dotenv==1.0.0 